fixes user-group delete method and adds user-group serializer

+++ app/controllers/user_groups_controller.rb
@@ -3,6 +3,11 @@ class UserGroupsController < ApplicationController
   # users can only join a group when logged in.
   before_action :authenticate_user
 
+  def index
+    user_groups = UserGroup.all
+    render json: user_groups
+  end
+
   def create
     user_group = UserGroup.new(
       user_id: current_user.id,
@@ -18,7 +23,11 @@ class UserGroupsController < ApplicationController
   #users can only leave a group when they are logged in
   def destroy 
     user_group = UserGroup.find(params[:id])
-    user_group.destroy
-    render json: {message: "You have successfully left the group"}
+    if user_group.user_id == current_user.id
+      user_group.destroy
+      render json: {message: "You have successfully left the group"}
+    else
+      render json: {message: "You do not have access to this account"}
+    end
   end
 end
